+++ includes/frontend/class-menu.php
@@ -6,7 +6,7 @@
  * @copyright  WebMan Design, Oliver Juhas
  *
  * @since    2.0.0
- * @version  2.6.1
+ * @version  2.6.3
  *
  * Contents:
  *
@@ -421,7 +421,7 @@ class Modern_Menu {
 			 * Social links supported icons
 			 *
 			 * @since    2.0.0
-			 * @version  2.6.1
+			 * @version  2.6.3
 			 */
 			public static function social_links_icons() {
 
@@ -474,6 +474,7 @@ class Modern_Menu {
 						'wa.me'             => 'whatsapp',
 						'wordpress.org'     => 'wordpress',
 						'wordpress.com'     => 'wordpress',
+						'x.com'          => 'x',
 						'xing.com'          => 'xing',
 						'yelp.com'          => 'yelp',
 						'youtube.com'       => 'youtube',
@@ -489,7 +490,7 @@ class Modern_Menu {
 			 * Note that the menu has to be set to output `<!--{{icon}}-->` placeholders!
 			 *
 			 * @since    2.0.0
-			 * @version  2.2.3
+			 * @version  2.6.3
 			 *
 			 * @param  string  $item_output The menu item output.
 			 * @param  WP_Post $item        Menu item object.
@@ -513,10 +514,32 @@ class Modern_Menu {
 
 				// Processing
 
-					foreach ( $social_icons as $url => $icon ) {
-						if ( false !== strpos( $item_output, $url ) ) {
-							$social_icon = $icon;
-							break;
+					if (
+						! empty( $item->classes )
+						&& false !== stripos( implode( ' ', (array) $item->classes ), 'has-icon-' )
+					) {
+
+						$forced_icon = array_intersect(
+							$social_icons,
+							array_map(
+								function( $item ) {
+									return str_replace( 'has-icon-', '', trim( $item ) );
+								},
+								(array) $item->classes
+							)
+						);
+
+						if ( ! empty( $forced_icon ) ) {
+							$social_icon = reset( $forced_icon );
+						}
+
+					} else {
+
+						foreach ( $social_icons as $url => $icon ) {
+							if ( false !== strpos( $item_output, $url ) ) {
+								$social_icon = $icon;
+								break;
+							}
 						}
 					}